[M42 Merge | Skia] Batch draw rectangle bitmap.
There are three draw bitmap paths in Skia,
1. draw bitmap via drawRect
2. draw bitmap via drawBitmapXXXX
3. draw bitmap via drawSprite

Draw bitmap call can be batched if two consequitive draws are of same
bitmap. For this, local rectangle of the bitmap should be set.
The local matrix of the bitmap is computed from the shader matrix
and view matrix. In case of drawBitmap and drawSprite, there is no
shader, so the shader matrix is identity matrix.

Test Platform: Ubuntu-x86/Mesa

+---------------------------+--------+--------+
|      Native Bitmap Test   | Before | After  |
+---------------------------+--------+--------+
| Performance score         |  10.5  |  11.4  |
+---------------------------+--------+--------+

M42 Patch: http://165.213.202.130/gerrit/#/c/81206/

Bug: http://suprem.sec.samsung.net/jira/browse/CBEFL-657
Reviewed by: DaeHyun Ko, venu musham

Change-Id: I6707a6b89bd37d68289d91e380c57875e494832d
Signed-off-by: Suyambulingam R M <suyambu.rm@samsung.com>