[M42 Merge | Skia] Create/reference a scratch texture from a bitmap only if it is not backed by a valid texture.
A bitmap can be backed by a valid texture. When converting
a bitmap to a GrFragmentProcessor from a bitmap shader, check whether
there is a valid texture associated with it. If not create/reference
a new texture. If there is a texture, we must check whether target
driver support npot and tiling. If not, we need to create a pot
texture and paint original texture to pot texture.

M42 Patch: http://165.213.202.130/gerrit/#/c/81194/

Bug: http://suprem.sec.samsung.net/jira/browse/CBEFL-657
Reviewed by: DaeHyun Ko, venu musham

Change-Id: Ibf1868ecb326a405844240bd7ef81b8bf68023c7
Signed-off-by: Suyambulingam R M <suyambu.rm@samsung.com>