Support emulator and refactor build configurations
1) Support emulator build
Gyp switches for arm architcture are removed in case of
emulator architecture build.

2) Refactor ninja
The home-built ninja binaries have been directly called in spec file.
This CL refactors ninja script to call the binaries in it.
The ninja binary file names are changed like below.
>> ninja-{OS}{LONG_BIT}.{ARCHITECTURE}

3) Change default build profiles
The tv, mobile, wearable targets are changed with proper ones.

4) Remove git
Requiring unused git binary is removed in spec file.
It used to clone source using submodules, but not now.

5) Update gbs.conf
Some of repository is updated with the latest one.

6) Update .gitignore
The output files of patche set are ignored.

7) Move pkgconfig and manifest
The skia.pc and skia.manifest are moved to proper path.

Bug: http://suprem.sec.samsung.net/jira/browse/TWF-1297

Change-Id: I0a9d607b73cf992ac5ffd2b520a3f3d8181a5dd7
Signed-off-by: Youngsoo Choi <kenshin.choi@samsung.com>