Refactor build configurations
1) Refactor ninja
The home-built ninja binaries have been directly called in spec file.
This CL refactors ninja script to call the binaries in it

2) Remove git
Unused git binary is removed in spec file.

3) Update gbs.conf
Some repositories are updated with the latest one.

4) Update .gitignore
The output files of patch set are ignored.

5) Move pkgconfig and manifest
The skia.pc and skia.manifest are moved to proper path.

Bug: http://suprem.sec.samsung.net/jira/browse/TWF-1297

Change-Id: I0a9d607b73cf992ac5ffd2b520a3f3d8181a5dd7
Signed-off-by: Youngsoo Choi <kenshin.choi@samsung.com>